Removal of CDW stored procedure support for DiagnosticReport. Test refactoring in progress.
